Python/requirements.txt
dash==2.18.2
plotly==5.24.1
pyserial==3.5
